requirements.txt
Flask==3.0.3
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
python-dotenv==1.0.1
marshmallow==3.21.3
passlib==1.7.4
bcrypt==4.2.0

